Fix updating tag counter
Property preference has an association with an old version of the tag
given bidirectional navigability and immutability requiring recreation
of tags. Deleting a property preferences updates the tag with the old
tag which does not contain the associations to property preferences,
thus outputting 0 as the tag counter when a property preference is
deleted.